Removed includes and libraries folders - this helps encourage the B.Y.O.B. mindset and hopefully prevents possible issues with licensing.
Added cmake folder and the associated Find*.cmake files needed for the framework.

Updated all headers that use GLEW to include "GL/glew.h" as opposed to "glew.h" to reflect what is returned by FindGLEW and where GLEW should be/is installed using make.

Removed Lua library, headers,associated code in Game.cpp, and the associated file being used in /assets. At this time, Lua is still being considered for use with the framework.

Tested building on Crunchbang GNU/Linux (waldorf). Builds successfully without errors, but appears to need the various .so's in the same directory to run the application.