(Yet Another) Clean-Up and Performance Pass
Major Changes Made Include:

-Finally using the Shader struct in the code, and have removed all
instances of the old shaderProgram/shaderPairMap paradigm.

-Due to the shift over to using the Shader struct, GameObjects had an
issue where they would need the shader to work with their Mesh's VAO; a
helper method was added to alleviate this issue for now.

-GameObject's no longer Draw() themselves; a Game::Draw() loop has been
added to iterate over the display list and render each object in the list
based on the data referred to by the given GameObject.

-The GameEntity class was renamed to Transform, and previous instances of
inheritting from GameEntity/Transform (outside of the Camera class,
temporarily) have been removed. Instead, classes that previously inherited
Transform now contain a Transform member variable.

-The ResourceManager has been updated to handle the creation of all objects
needed in a Game (i.e. Cameras, Shaders, Lights, GameObjects, and Meshes
thus far). When created, this data is store in vectors, and a pointer is
returned. Creation of game elements should only be done via the
ResourceManager.

-OpenAL stub code has been removed, as it is out of scope for the near
future, and other options such as Ogg Vorbis are being considered.

-The std::maps in ResourceManager have been changed to std::unordered_maps
with the idea that once all of the data for a given level is loaded, the
data loaded won't be changed frequently or significantly, and therefore
the look-up speed of the underlying hashtable should be faster.

-In Game, the returned value of GLFW's getTime() function is cast to a
float value for use with/as deltaTime. This was done to prevent recasting
the value any time that deltaTime would be used, specifically within
Game::Update when applying it to every object within the display list.

-Misc. code clean up, minor optimizations, and documentation.